requirements.txt
streamlit==1.33.0
python-dotenv==1.0.0
google-generativeai==0.3.2  # Untuk integrasi Gemini